moved tests to google/drive directory
