+++ src/appmixer/google/drive/artifacts/tests/normalization.test.js
